+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/ModuleConfiguration.cs
@@ -2,8 +2,10 @@ namespace L2Data2Code.BaseGenerator.Configuration
 {
     public class ModuleConfiguration
     {
+        public string Key { get; set; }
         public string Name { get; set; }
         public string IncludeTables { get; set; }
         public string ExcludeTables { get; set; }
+        public string Group { get; set; }
     }
 }