+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/ModulesConfiguration.cs
@@ -7,6 +7,27 @@ namespace L2Data2Code.BaseGenerator.Configuration
     {
         public ModulesConfiguration(IJsonSetting jsonSetting) : base(jsonSetting, GeneratorSectionLabels.MODULES)
         {
+            SetVars();
+            jsonSetting.PropertyChanged += (s, e) =>
+            {
+                if (e.PropertyName == nameof(JsonSetting.Config))
+                {
+                    SetVars();
+                }
+            };
         }
+
+        private void SetVars()
+        {
+            foreach (var key in GetKeys())
+            {
+                this[key].Key = key;
+                if (this[key].Group == null)
+                {
+                    this[key].Group = key;
+                }
+            }
+        }
+
     }
 }