+++ src/L2Data2Code/L2Data2Code.SchemaReader/Configuration/DataSourcesConfiguration.cs
@@ -9,6 +9,25 @@ namespace L2Data2Code.BaseGenerator.Configuration
         private const string DEFAULT_KEY = "localserver";
         public DataSourcesConfiguration(IJsonSetting jsonSetting) : base(jsonSetting, ConfigurationLabels.DATA_SOURCES)
         {
+            SetVars();
+            jsonSetting.PropertyChanged += (s, e) =>
+            {
+                if (e.PropertyName == nameof(JsonSetting.Config))
+                {
+                    SetVars();
+                }
+            };
+        }
+
+        private void SetVars()
+        {
+            foreach (var key in GetKeys())
+            {
+                if (this[key].ModulesGroup == null)
+                {
+                    this[key].ModulesGroup = this[key].DefaultModule;
+                }
+            }
         }
 
         public string Schema(string key)