+++ src/L2Data2Code/L2Data2Code.Avalonia/changes.md
@@ -1,3 +1,9 @@
+### 3.2.4
+
+- Posibility to group name spaces using module groups
+- **New**: Added `ModulesGroup` to `DataSources`
+- **New**: Added `Group` to `Modules`
+
 ### 3.2.3
 
 - Upgrade Material.Icons.Avalonia to 2.1.0